@import: dependency was not found
I'm using webpack and the modern-normalize.css was not found with the import path provided in the readme. I had to use the tilde character to point to the node_modules folder. I thought it would be nice to share this tip with others.

Thanks for this normalizing style-sheet. It's perfect for projects where I don't care about non evergreen browsers 👍